Moved display animation handler
